move to computed gatherer
