+++ lighthouse-core/audits/dobetterweb/uses-optimized-images.js
@@ -27,7 +27,6 @@
 const Audit = require('../audit');
 const URL = require('../../lib/url-shim');
 const Formatter = require('../../formatters/formatter');
-const NetworkHelper = require('../../lib/network-recorder');
 
 const KB_IN_BYTES = 1024;
 const IGNORE_THRESHOLD_IN_BYTES = 2 * KB_IN_BYTES;
@@ -68,9 +67,19 @@ class UsesOptimizedImages extends Audit {
    * @return {!AuditResult}
    */
   static audit(artifacts) {
-    const images = artifacts.OptimizedImages;
     const networkRecords = artifacts.networkRecords[Audit.DEFAULT_PASS];
-    const networkThroughput = NetworkHelper.computeAverageThroughput(networkRecords);
+    return artifacts.requestNetworkThroughput(networkRecords).then(networkThroughput => {
+      return UsesOptimizedImages.audit_(artifacts, networkThroughput);
+    });
+  }
+
+  /**
+   * @param {!Artifacts} artifacts
+   * @param {number} networkThroughput
+   * @return {!AuditResult}
+   */
+  static audit_(artifacts, networkThroughput) {
+    const images = artifacts.OptimizedImages;
 
     if (images.rawValue === -1) {
       return UsesOptimizedImages.generateAuditResult(images);
