+++ lighthouse-core/gather/computed/network-throughput.js
@@ -1,0 +1,72 @@
+/**
+ * @license
+ * Copyright 2017 Google Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+'use strict';
+
+const ComputedArtifact = require('./computed-artifact');
+
+class NetworkThroughput extends ComputedArtifact {
+  get name() {
+    return 'NetworkThroughput';
+  }
+
+  /**
+   * Computes the average throughput for the given records in bytes/second.
+   * Excludes data URI, failed or otherwise incomplete, and cached requests.
+   *
+   * @param {?Array<!WebInspector.NetworkRequest>} networkRecords
+   * @return {number}
+   */
+  compute_(networkRecords) {
+    if (!networkRecords || !networkRecords.length) {
+      return 0;
+    }
+
+    let totalBytes = 0;
+    const timeBoundaries = networkRecords.reduce((boundaries, record) => {
+      if (/^data:/.test(record.url) || record.failed || !record.finished ||
+          record.statusCode > 300 || !record.transferSize) {
+        return boundaries;
+      }
+
+      totalBytes += record.transferSize;
+      boundaries.push({time: record.responseReceivedTime, isStart: true});
+      boundaries.push({time: record.endTime, isStart: false});
+      return boundaries;
+    }, []).sort((a, b) => a.time - b.time);
+
+    let inflight = 0;
+    let currentStart = 0;
+    let totalDuration = 0;
+    timeBoundaries.forEach(boundary => {
+      if (boundary.isStart) {
+        if (inflight === 0) {
+          currentStart = boundary.time;
+        }
+        inflight++;
+      } else {
+        inflight--;
+        if (inflight === 0) {
+          totalDuration += boundary.time - currentStart;
+        }
+      }
+    });
+
+    return totalBytes / totalDuration;
+  }
+}
+
+module.exports = NetworkThroughput;
